refactors helper functions and config
Config is now loaded from ~/.herdconfig or the envar "herd_config_path"
It's an ini file that currently needs a [Build] heading with base_path
and host set.
